Need to work on viewmodels to get specific elements.

+++ app/src/main/java/com/example/assignment3/repositories/BoardGamesRepository.kt
@@ -33,7 +33,7 @@ class BoardGamesRepository(
         _BoardGames.value += newBoardGame
     }
 
-    suspend fun getBoardGameById(id: Int): BoardGame? {
+    suspend fun getBoardGameById(id: Long): BoardGame? {
         return boardGamesDao.getBoardGameById(id)
     }
